fix: adjust release download link
